// src/types/serie.ts
export interface Serie {
  id: string;
  name: string;
  description: string;
  tagline: string;
  image: string;
  imagelandscape: string;
  status: string;
  read: string;
  legal: string;
  credits: string;
  comments: string;
  genre: string;
  clasification: string;
  country: string;
  language: string;
  total_chapters: number;
  likes: number;
  bookmarks: number;
}